# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tutor/sessions:
    get:
      operationId: listSessions
      tags: [tutor]
      summary: List chat sessions for a device
      parameters:
        - name: deviceId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
    post:
      operationId: createSession
      tags: [tutor]
      summary: Create a new chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionBody"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /tutor/sessions/{id}:
    get:
      operationId: getSession
      tags: [tutor]
      summary: Get a session with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSessionWithMessages"
        "404":
          description: Not found
    delete:
      operationId: deleteSession
      tags: [tutor]
      summary: Delete a chat session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tutor/sessions/{id}/messages:
    post:
      operationId: sendMessage
      tags: [tutor]
      summary: Send a message to the AI tutor (streaming SSE)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "200":
          description: SSE stream of AI tutor response chunks
          content:
            text/event-stream: {}
  /tutor/sessions/{id}/image-messages:
    post:
      operationId: sendImageMessage
      tags: [tutor]
      summary: Send an image (OCR) message to the AI tutor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendImageMessageBody"
      responses:
        "200":
          description: SSE stream of AI tutor response for image question
          content:
            text/event-stream: {}
...
components:
  schemas:
...
    ChatSession:
      type: object
      properties:
        id:
          type: integer
        deviceId:
          type: string
        subject:
          type: string
        mode:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - deviceId
        - subject
        - mode
        - title
        - createdAt
        - updatedAt
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - role
        - content
        - createdAt
    ChatSessionWithMessages:
      type: object
      properties:
        id:
          type: integer
        deviceId:
          type: string
        subject:
          type: string
        mode:
          type: string
        title:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - deviceId
        - subject
        - mode
        - title
        - messages
        - createdAt
        - updatedAt
    CreateSessionBody:
      type: object
      properties:
        deviceId:
          type: string
        subject:
          type: string
        mode:
          type: string
        title:
          type: string
        grade:
          type: string
        board:
          type: string
      required:
        - deviceId
        - subject
        - mode
        - title
        - grade
        - board
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
        deviceId:
          type: string
        grade:
          type: string
        board:
          type: string
        subject:
          type: string
        mode:
          type: string
      required:
        - content
        - deviceId
        - grade
        - board
        - subject
        - mode
    SendImageMessageBody:
      type: object
      properties:
        imageBase64:
          type: string
          description: Base64 encoded image of handwritten or printed question
        deviceId:
          type: string
        grade:
          type: string
        board:
          type: string
        subject:
          type: string
        mode:
          type: string
      required:
        - imageBase64
        - deviceId
        - grade
        - board
        - subject
        - mode
...