# lib/api-zod/src/generated/types/chatSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatSession {
  id: number;
  deviceId: string;
  subject: string;
  mode: string;
  title: string;
  createdAt: Date;
  updatedAt: Date;
}