# lib/api-zod/src/generated/types/sendMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SendMessageBody {
  content: string;
  deviceId: string;
  grade: string;
  board: string;
  subject: string;
  mode: string;
}